Adds head_js_analytics block and analytics_domain
Adds a head_js_analytics block that opens immediately after the head
tag to allow running analytic head scripts such as chartbeat.

Adds in a check for analytics_domain in the google analytics section
and if found, we push _setDomainName onto _gaq.

Updates composer.json to require symfony 2.1 at minimum to allow for
symfony 2.3 and on.